nick: the order by clause for the front page of the gallery is set as: coalesce(PinholePhoto.photo_date, PinholePhoto.publish_date) desc, id desc
do we really want to order by photo_date first? The current gallery only orders
by publish_date so that it's the newest photos added that are displayed, not
necessarily the date they were taken.

steven: hrm... browsing by photo date makes the most sense, but for the home page, we'd
probably want publish date

nathan: if we do use photo_date, i'll add time zones to that...

dan: If I add photos of me as a kid I would like them to show up on the top of the
reverse chronological order, not automatically fall back to 1983

steven: right

nick: exactly.

dan: agreed

nick: we could have a link that was like "order chronologically" that would add site.orderby=photo-date or something.



git-svn-id: file:///Users/gauthierm/svnrepo/pinhole/trunk@19029 94490950-a86c-4493-9ccb-530f72933e47