Add smarter display to the number of atom posts. As it states in the docs:
Photos are displayed reverse-chronologically based on their publish-date.
The number of photos is always at least $min_entries, but if a recently
published batch of photos (within the time of $recent_period) exceeds
$min_entries, up to $max_entries photos will be displayed. This makes it
easier to ensure that a subscriber won't miss part of a batch, while
limiting server load for the feed.



git-svn-id: file:///Users/gauthierm/svnrepo/public/pinhole/trunk@32770 94490950-a86c-4493-9ccb-530f72933e47